@@ -11,38 +11,42 @@
     "format": "prettier --write ."
   },
   "dependencies": {
-    "convict": "^6.0.1",
-    "convict-format-with-validator": "^6.0.1",
+    "axios": "1.6.2",
+    "convict": "^6.2.4",
+    "convict-format-with-validator": "^6.2.0",
     "cors": "^2.8.5",
-    "dotenv": "^16.0.1",
-    "express": "^4.18.1",
-    "helmet": "^5.1.0",
-    "mongoose": "^6.4.3",
-    "socket.io": "^4.0.1",
-    "uuid": "^8.3.2"
+    "dotenv": "^16.3.1",
+    "express": "^4.18.2",
+    "helmet": "^7.1.0",
+    "mongoose": "^8.0.3",
+    "react-cosmos": "^5.7.2",
+    "react-scripts": "^5.0.1",
+    "socket.io": "^4.7.2",
+    "uuid": "^9.0.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.18.6",
-    "@babel/preset-env": "^7.18.6",
-    "@babel/preset-typescript": "^7.18.6",
-    "@shelf/jest-mongodb": "^3.0.2",
-    "@types/convict": "^6.0.1",
-    "@types/convict-format-with-validator": "^6.0.2",
-    "@types/cors": "^2.8.10",
-    "@types/express": "^4.17.13",
-    "@types/jest": "^28.1.4",
-    "@types/node": "^18.0.2",
-    "@types/socket.io": "^2.1.13",
-    "@types/uuid": "^8.3.4",
-    "@typescript-eslint/eslint-plugin": "^5.30.5",
-    "@typescript-eslint/parser": "^5.30.5",
-    "babel-jest": "^28.1.2",
-    "eslint": "^8.19.0",
-    "eslint-config-prettier": "^8.3.0",
-    "jest": "^28.1.2",
-    "nodemon": "^2.0.19",
-    "prettier": "^2.7.1",
-    "ts-node": "^10.8.2",
-    "typescript": "^4.7.4"
+    "@babel/core": "^7.23.6",
+    "@babel/preset-env": "^7.23.6",
+    "@babel/preset-typescript": "^7.23.3",
+    "@shelf/jest-mongodb": "^4.2.0",
+    "@types/convict": "^6.1.6",
+    "@types/convict-format-with-validator": "^6.0.5",
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/jest": "^29.5.11",
+    "@types/node": "^20.10.5",
+    "@types/socket.io": "^3.0.1",
+    "@types/uuid": "^9.0.7",
+    "@typescript-eslint/eslint-plugin": "^6.15.0",
+    "@typescript-eslint/parser": "^6.15.0",
+    "babel-jest": "^29.7.0",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.2",
+    "jest": "^29.7.0",
+    "nodemon": "^3.0.2",
+    "prettier": "^3.1.1",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3"
   }
 }
