@@ -14,44 +14,44 @@
   },
   "proxy": "http://localhost:42069",
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.35",
-    "@fortawesome/free-solid-svg-icons": "^5.15.3",
-    "@fortawesome/react-fontawesome": "^0.1.14",
-    "@testing-library/jest-dom": "^5.12.0",
-    "@testing-library/react": "^11.2.6",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^28.1.4",
-    "@types/node": "^18.0.3",
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.3",
-    "axios": "^0.27.2",
-    "history": "^4.10.1",
+    "@fortawesome/fontawesome-svg-core": "^6.5.1",
+    "@fortawesome/free-solid-svg-icons": "^6.5.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@testing-library/jest-dom": "^6.1.5",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/user-event": "^14.5.1",
+    "@types/jest": "^29.5.11",
+    "@types/node": "^20.10.5",
+    "@types/react": "^18.2.45",
+    "@types/react-dom": "^18.2.18",
+    "axios": "^1.6.2",
+    "history": "^5.3.0",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0",
+    "nodemon": "3.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.21.1",
     "react-scripts": "^5.0.1",
-    "socket.io-client": "^4.0.1",
-    "typescript": "^4.7.4",
+    "socket.io-client": "^4.7.2",
+    "typescript": "^5.3.3",
     "use-persisted-state": "^0.3.3",
-    "web-vitals": "^1.1.1"
+    "web-vitals": "^3.5.0"
   },
   "devDependencies": {
-    "@testing-library/react-hooks": "^5.1.2",
-    "@types/lodash": "^4.14.168",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/react-test-renderer": "^17.0.1",
-    "@types/use-persisted-state": "^0.3.0",
-    "@typescript-eslint/eslint-plugin": "^5.30.5",
-    "@typescript-eslint/parser": "^5.30.5",
-    "eslint": "^8.19.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-react": "^7.30.1",
+    "@types/lodash": "^4.14.202",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-test-renderer": "^18.0.7",
+    "@types/use-persisted-state": "^0.3.4",
+    "@typescript-eslint/eslint-plugin": "^6.15.0",
+    "@typescript-eslint/parser": "^6.15.0",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "jest-mock-axios": "^4.6.1",
-    "prettier": "^2.7.1",
-    "react-cosmos": "^5.6.2",
-    "react-test-renderer": "^17.0.2"
+    "jest-mock-axios": "^4.7.3",
+    "prettier": "^3.1.1",
+    "react-cosmos": "^5.7.2",
+    "react-test-renderer": "^18.2.0"
   },
   "browserslist": {
     "production": [
