@@ -1,13 +1,15 @@
 import React from "react";
-import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { PunchlinesProvider } from "./contexts/punchlines";
 import { RoundProvider } from "./contexts/round";
 import { PlayersProvider } from "./contexts/players";
+import { createRoot } from "react-dom/client";
 
-ReactDOM.render(
+const container = document.getElementById("root");
+const root = createRoot(container!);
+root.render(
   <React.StrictMode>
     <PlayersProvider>
       <PunchlinesProvider>
@@ -16,8 +18,7 @@ ReactDOM.render(
         </RoundProvider>
       </PunchlinesProvider>
     </PlayersProvider>
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
